Fixed issues in useAuth.tsx and main.tsx
