+++ src/api/types/env.d.ts
@@ -7,15 +7,4 @@ declare namespace NodeJS {
 
       // Add other env vars like PORT, NODE_ENV if needed
     }
-
-    interface ImportMetaEnv {
-      readonly VITE_BACKEND_API_URL_LOCAL: string;
-      // Add other environment variables here
-      // readonly VITE_API_KEY: string;
-    }
-
-    interface ImportMeta {
-      readonly env: ImportMetaEnv;
-    }
-
 }
