+++ src/client/main.tsx
@@ -1,29 +1,19 @@
-import React from 'react';
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import { RecoilRoot } from 'recoil';
-import './styles/globals.css'
+import React, { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './App';
+import './styles/globals.css';
 
-import { BrowserRouter, Routes, Route, useNavigate } from 'react-router-dom';
-import LoginPage from './pages/Login/loginpage.jsx';
-import { ThemeProvider } from './helper/dark-mode-provider/theme-provider';
-import PrivateRoute from './helper/private-route/private-route.jsx';
-import DashboardPage from './pages/Dashboard/DashboardPage.jsx';
+// More defensive approach with type checking
+const rootElement = document.getElementById('root');
 
-createRoot(document.getElementById('root')).render(
+if (!rootElement) {
+  throw new Error('Failed to find the root element');
+}
+
+const root = createRoot(rootElement);
+
+root.render(
   <StrictMode>
-    <RecoilRoot>
-        <ThemeProvider defaultTheme='light' storageKey='vite-ui-theme'>
-          <BrowserRouter>
-            <Routes>
-              <Route element={<PrivateRoute/>}>
-                <Route path="/" element={<DashboardPage/>} />
-              </Route>
-
-              <Route path="/login" element={<LoginPage />} />
-            </Routes>
-          </BrowserRouter>
-        </ThemeProvider>
-      </RecoilRoot>
-  </StrictMode>,
-)
+    <App />
+  </StrictMode>
+);
