+++ src/client/vite-env.d.ts
@@ -1,0 +1,5 @@
+interface ImportMeta {
+  readonly env: {
+      VITE_BACKEND_API_URL_LOCAL: string;
+  }
+}
