+++ tsconfig.server.json
@@ -3,6 +3,7 @@
       "target": "ESNext",
       "module": "NodeNext",
       "moduleResolution": "NodeNext",
+      "jsx": "react-jsx",
       "strict": false,
       "esModuleInterop": true,
       "skipLibCheck": true,
